Extract cache update logic into shared helper module
- Create nym-vpn-lib/cache_refresh module
- Ensure gateway and topology caches are properly invalidated on discovery updates
